clients: Handle nonexistent accounts in rucio-admin account get-limits fixes #7458
Previously, `rucio-admin account get-limits` accepted nonexistent accounts and returned misleading responses. This fix adds a validation check to ensure that the account exists before retrieving limits. If the account does not exist, an error exception is raised, and the command returns a failure. This also adds error handling while retrieving limits of accounts.

Signed-off-by: Karanjot Singh <karanjot.singh@cern.ch>